refactor: add index entry and rename to useCards
#39
